backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
youtube-transcript-api
openai-whisper
bert-extractive-summarizer
nltk
moviepy